% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oemdlRtools.R
\name{diffp}
\alias{diffp}
\title{Period-on-period differences}
\usage{
diffp(x)
}
\arguments{
\item{x}{A numeric vector containg the original values.}
}
\value{
A numeric vector containing the transformed values.
}
\description{
A convenience function to calculate period-on-period differences for a
given data series. If the input vector is of annual frequency, the function
will return year-on-year differences. If the input vector is of quarterly
frequency, the function will return quarter-on-quarter differences. Note
that the frequency of the input vector is not checked for explicitly.
}
\examples{
diffp(1:10)
}
