% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oemdlRtools.R
\name{oe_macromappings}
\alias{oe_macromappings}
\title{Return Oxford Economics sector names and corresponding ISO codes}
\usage{
oe_macromappings()
}
\value{
Data frame containing Oxford Economics sector names and corresponding
ISO 3-character code.
}
\description{
Returns a data frame containing Oxford Economics sector names and
corresponding ISO 3-character code.
}
\examples{
oe_macromappings()
}
