% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oemdlRtools.R
\name{oe_pal}
\alias{oe_pal}
\title{Function to return Oxford Economics colour palette (discrete)}
\usage{
oe_pal()
}
\value{
Returns Oxford Economics colour palette.
}
\description{
Returns Oxford Economics colour palette. This palette includes 15 unique
colours.
}
\examples{
oe_pal()
}
