% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oemdlRtools.R
\name{oe_palette}
\alias{oe_palette}
\title{Return Oxford Economics colour palette}
\usage{
oe_palette()
}
\value{
Returns Oxford Economics colour palette.
}
\description{
Returns Oxford Economics colour palette. The palette includes 15 unique
colours.
}
\examples{
oe_palette()
}
