% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oemdlRtools.R
\name{pach}
\alias{pach}
\title{Annualised quarterly growth rates}
\usage{
pach(x)
}
\arguments{
\item{x}{A numeric vector containg the original values.}
}
\value{
A numeric vector containing the transformed values.
rates.
}
\description{
A convenience function to calculate quarter-on-quarter annualised growth
rates for a given data series. Note that frequency of the input vector is not
checked for explicitly.
}
\examples{
pach(1:10)
}
