% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oemdlRtools.R
\name{pch}
\alias{pch}
\title{Period-on-period growth rates}
\usage{
pch(x)
}
\arguments{
\item{x}{A numeric vector containg the original values.}
}
\value{
A numeric vector containing the transformed values.
}
\description{
A convenience function to calculate period-on-period growth rates for a
given data series. If the input vector is of annual frequency, the function
will return year-on-year growth rates. If the input vector is of quarterly
frequency, the function will return quarter-on-quarter growth rates. Note
that the frequency of the input vector is not checked for explicitly.
}
\examples{
pch(1:10)
}
