% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oemdlRtools.R
\name{pchy}
\alias{pchy}
\title{Year-on-year growth rates}
\usage{
pchy(x)
}
\arguments{
\item{x}{A numeric vector containg the original values.}
}
\value{
A numeric vector containing the transformed values.
}
\description{
A convenience function to calculate year-on-year growth rates for a given
data series. Note that frequency of the input vector is not checked for
explicitly.
}
\examples{
pchy(1:20)
}
