% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oemdlRtools.R
\name{read_oedb}
\alias{read_oedb}
\title{Read Oxford Economics Global Economic Model database files}
\usage{
read_oedb(db, md = "C:/OEF", sy = 1980, ey = 2050, id = NULL,
  tp = "V")
}
\arguments{
\item{db}{A filename.}

\item{md}{Model directory (default: C:/OEF).}

\item{sy}{The first year for which data should be exported (default: 1980).}

\item{ey}{The last year for which data should be exported (default: 2050).}

\item{id}{A vector of variable identifiers to be exported (optional).
Variable identifiers are made up of variable names and location names,
separated by a comma. If omitted, all variables will be imported.}

\item{tp}{Type of values to be exported: (\code{V}) for variable data
(default), \code{R} for residual data.}
}
\value{
A list containing the data frame (\code{$df}), basic meta data
(\code{$meta}), and a vector of dates giving the last data point with
historical data (\code{$lasthistory})
}
\description{
Reads an Oxford Economics Global Economic Model database file (.db) into a
data frame.
}
\examples{
setwd("C:/OEF/")
ids <- c("GDPHEADPPP!,US", "GDPHEADPPP!,GERMANY", "GDPHEADPPP!,UK",
          "GDPHEADPPP!,JAPAN", "GDPHEADPPP!,KOREA")
dat <- read_oedb("Jan19.db", "C:/OEF", 1990, 2020, ids, "V")$df
head(dat)
}
\seealso{
\code{\link{oe_macromappings}}
}
