% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oemdlRtools.R
\name{read_oedb}
\alias{read_oedb}
\title{Read Oxford Economics Global Economic Model database files}
\usage{
read_oedb(
  db,
  mnemonic = NULL,
  sector = NULL,
  mnemonic_sector = NULL,
  exp_type = "V",
  model_dir = "C:/OEF",
  start_year = 1980,
  end_year = 2050,
  as_xts = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{db}{Character. Filename of or path to the database file.}

\item{mnemonic}{Character or character vector. Mnemonics (variable names) to 
import.}

\item{sector}{Character or character vector. Sectors (country names) to
import.}

\item{mnemonic_sector}{Data frame with two columns, "Mnemonic" and "Sector".
These specify custom mnemonic-sector combinations.}

\item{exp_type}{Character or character vector. Type of values to be imported:
\code{"V"} for variable data (default), \code{"R"} for residual data, or
\code{c("V", "R")} for both.}

\item{model_dir}{Character. Path to model directory (default: C:/OEF).}

\item{start_year}{Numeric. The first year for which to import data (default:
1980).}

\item{end_year}{Numeric. The last year for which to import data (default:
2050).}

\item{as_xts}{Logical. If \code{TRUE}, data is returned in xts format.}

\item{verbose}{Logical. If \code{TRUE}, status messages are printed.}
}
\value{
A list containing the data \code{$dat}, header metadata
\code{$dat_head}, fix metadata \code{$fix}, variable metadata \code{$var}
(including a logical value \code{$var$Is.percent} indicating if a variable is
a percentage value, and a character value \code{$var$Residual.Indicator}
listing the residual name corresponding to a specific variable), a character
vector \code{$type} storing the type of data that has been imported (variable
or residual data), and numeric values \code{$last_hist} and
\code{$first_fcst} specifying the index values of the last historical data
point and the first forecast data point respectively. Note that mnemonics are
converted to syntactically valid R names and that an "R_" prefix is added to
names of columns containing residual data.
}
\description{
Imports data from an Oxford Economics Global Economic Model database file
(.db) into R.
}
\seealso{
\code{\link{oe_macromappings}}
}
